Add universal league page
The Dashboard page is currently under construction and being adjusted to follow a more consistent UI them.

What remains is a better lookingleague summary as well as a better looking structurew for displaying information on the individual league page.

The universal league page has also been added and still needs to be adjust ed to look different when compared to the individual league page.